chore: refactor & clean up snapshots
