+++ packages/provider/src/provider.js
@@ -1,5 +1,4 @@
-import connect from "./connect";
-import QueryProvider from "./query-provider";
+import connect, { QueryProvider } from "./connect";
 import AuthorProfileProvider from "./author-profile";
 import AuthorArticlesNoImagesProvider from "./author-articles-no-images";
 import AuthorArticlesWithImagesProvider from "./author-articles-with-images";
